requirements.txt
fastapi
uvicorn
pymongo
pydantic
bcrypt
python-multipart
google-generativeai
pdfplumber
langchain
langchain-text-splitters
langchain-google-genai
pydantic-settings
passlib
python-jose
python-dotenv
transformers
torch
sentence-transformers
SentencePiece
accelerate
llama-cpp-python
python-docx
motor
openpyxl
pandas
tenacity
pymongo
pytesseract
asyncpg
psycopg2-binary
latex2mathml

